requirements.txt
certifi
cffi
charset-normalizer
click
colorama
contourpy
cryptography
cycler
Flask
Flask-Login
Flask-SQLAlchemy
fonttools
greenlet
gunicorn
idna
itsdangerous
Jinja2
kiwisolver
MarkupSafe
matplotlib
numpy
objectpath
packaging
pandas
Pillow
pycparser
pyOpenSSL
pyparsing
python-dateutil
pytz
rauth
requests
scipy
six
SQLAlchemy
urllib3
Werkzeug
PyJWT
flask_migrate
